Eliminated redundant CSS and moved datafile
WeatherContainer has been swapped for generic Container
catalogsData.js moved to components folder so that src/ directory is less cluttered
