--- src/components/WeatherContainer.css
@@ -1,7 +1,0 @@
-.WeatherContainer {
-    width: 400px;
-    background-color: #1C1517;
-    border: 1px solid #ddd;
-    padding: 8px 8px;
-    text-align: center;
-}
