+++ src/index.css
@@ -16,3 +16,10 @@ code {
     monospace;
 }
 
+.Container {
+    width: 400px;
+    background-color: #1C1517;
+    border: 1px solid #ddd;
+    padding: 8px 8px;
+    text-align: center;
+}
